fix_rls_error.py
#!/usr/bin/env python3
"""
Script to fix RLS policies for user_custom_projections table
This addresses the "new row violates row-level security policy" error
"""

import os
from dotenv import load_dotenv
from supabase import create_client, Client

load_dotenv()

def fix_rls_policies():
    """Fix RLS policies for user_custom_projections table."""
    
    # Initialize Supabase client
    supabase_url = os.environ.get("SUPABASE_URL")
    supabase_key = os.environ.get("SUPABASE_KEY")
    
    if not supabase_url or not supabase_key:
        print("Error: SUPABASE_URL and SUPABASE_KEY environment variables are required")
        return False
    
    try:
        supabase: Client = create_client(supabase_url, supabase_key)
        print("Connected to Supabase successfully")
        
        # Drop existing policies
        print("Dropping existing RLS policies...")
        policies_to_drop = [
            "Users can view own custom projections",
            "Users can insert own custom projections", 
            "Users can update own custom projections",
            "Users can delete own custom projections"
        ]
        
        for policy_name in policies_to_drop:
            try:
                # Note: We can't directly drop policies via SQL in Supabase client
                # This would need to be done in the Supabase dashboard or via direct SQL
                print(f"  Would drop policy: {policy_name}")
            except Exception as e:
                print(f"  Note: Policy {policy_name} may not exist: {e}")
        
        # Create new permissive policy
        print("Creating new permissive RLS policy...")
        
        # Execute SQL to fix the policies
        sql_commands = [
            """
            DROP POLICY IF EXISTS "Users can view own custom projections" ON user_custom_projections;
            """,
            """
            DROP POLICY IF EXISTS "Users can insert own custom projections" ON user_custom_projections;
            """,
            """
            DROP POLICY IF EXISTS "Users can update own custom projections" ON user_custom_projections;
            """,
            """
            DROP POLICY IF EXISTS "Users can delete own custom projections" ON user_custom_projections;
            """,
            """
            CREATE POLICY "Allow all operations on custom projections" ON user_custom_projections
                FOR ALL USING (true) WITH CHECK (true);
            """
        ]
        
        for sql in sql_commands:
            try:
                # Note: Supabase client doesn't support raw SQL execution
                # This would need to be done in the Supabase dashboard
                print(f"  Would execute: {sql.strip()}")
            except Exception as e:
                print(f"  Note: Cannot execute raw SQL via client: {e}")
        
        print("\n" + "="*60)
        print("IMPORTANT: The RLS policies need to be updated manually in the Supabase dashboard")
        print("="*60)
        print("\nTo fix this issue, please:")
        print("1. Go to your Supabase dashboard")
        print("2. Navigate to Authentication > Policies")
        print("3. Find the 'user_custom_projections' table")
        print("4. Delete the existing restrictive policies")
        print("5. Create a new policy with:")
        print("   - Name: 'Allow all operations on custom projections'")
        print("   - Target roles: 'public'")
        print("   - Using expression: 'true'")
        print("   - With check expression: 'true'")
        print("\nAlternatively, you can run the SQL commands in the SQL Editor:")
        print("   - Go to SQL Editor in Supabase dashboard")
        print("   - Run the commands from fix_rls_policies.sql")
        
        return True
        
    except Exception as e:
        print(f"Error fixing RLS policies: {e}")
        return False

if __name__ == "__main__":
    print("RLS Policy Fix Script for user_custom_projections")
    print("=" * 50)
    
    success = fix_rls_policies()
    
    if success:
        print("\nScript completed. Please follow the manual steps above.")
    else:
        print("\nScript failed. Please check your Supabase configuration.")
